requirements.txt
# Core numerical + ML
numpy>=1.24
scipy>=1.10
torch>=2.1
h5py>=3.8
pyyaml

# Data structures
pyroaring>=0.3.6

# Server + serialization
flask>=2.3
msgpack>=1.0.5
waitress>=2.1

# Plotting (used by dataset_stats.py)
matplotlib>=3.7